+++ app/com/simplex9/splendor/Param.scala
@@ -4,10 +4,10 @@ package com.simplex9.splendor
   * Created by hongbo on 6/10/17.
   */
 object Param {
-  val WINNING_POINTS : Byte = 15
-  val NOBLE_POINT : Byte = 3
-  val NUM_COIN = Array[Byte](0,0,4,5,7)
-  val NUM_GOLD : Byte = 5
+  val WINNING_POINTS : Short = 15
+  val NOBLE_POINT : Short = 3
+  val NUM_COIN = Array[Short](0,0,4,5,7)
+  val NUM_GOLD : Short = 5
   val NUM_CARD_LEVEL = 3
   val NUM_CARD_EACH_LEVEL = 4
   val MAX_COIN = 10
